split apiTypes from client types
